- removed target_spell_effect_mod_flat and target_spell_effect_mod_pct * THERE IS A SIMPLE WAY TO FIX THESE! ADD A NEW SPELL GROUP FOR THOSE SPELLS! THERE IS A 64 BIT INTEGER RESERVED FOR THIS, I'M SURE THAT IS PLENTY. WE DON'T NEED TO WASTE MEMORY ON THIS CRAP! *
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4254 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Container.cpp
@@ -173,10 +173,10 @@ bool Container::AddItem(int8 slot, Item *item)
 void Container::SwapItems(int8 SrcSlot, int8 DstSlot)
 {
 	Item *temp;
-	if( SrcSlot < 0 || SrcSlot >= m_itemProto->ContainerSlots )
+	if( SrcSlot < 0 || SrcSlot >= (int8)m_itemProto->ContainerSlots )
 		return;
 
-	if( DstSlot < 0 || DstSlot >= m_itemProto->ContainerSlots )
+	if( DstSlot < 0 || DstSlot >= (int8)m_itemProto->ContainerSlots )
 		return;
 
 	if(m_Slot[DstSlot] &&  m_Slot[SrcSlot]&&m_Slot[DstSlot]->GetEntry()==m_Slot[SrcSlot]->GetEntry() && m_Slot[SrcSlot]->wrapped_item_id == 0 && m_Slot[DstSlot]->wrapped_item_id == 0 && m_Slot[DstSlot]->GetProto()->MaxCount>1)
